Refactoring to match cbioportal-frontend file structure
- Expanded documentation in README
- Refactoring to match the [cbioportal-frontend](https://github.com/cBioPortal/cbioportal-frontend) file structure
- Set explicit versions of node/yarn for running the dashboard
