+++ package.json
@@ -1,22 +1,41 @@
 {
-  "name": "smile-dashboard",
+  "name": "smile-dashboard-dev",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@babel/core": "7.4.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "8.0.5",
+    "babel-plugin-transform-es2015-modules-umd": "^6.22.0",
+    "babel-polyfill": "^6.22.0",
+    "babel-register": "^6.22.0",
+    "mobx": "^6.0.0",
+    "mobx-logger": "^0.7.1",
+    "mobx-react": "6.0.0",
+    "mobx-react-devtools": "6.1.1",
+    "mobx-react-lite": "3.0.1",
+    "mobx-react-router": "4.1.0",
+    "mobx-utils": "6.0.1",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "add": "^2.0.6",
-    "mobx": "^6.5.0",
-    "mobx-react": "^7.4.0",
-    "react": "^18.1.0",
-    "react-bootstrap": "^2.4.0",
-    "react-dom": "^18.1.0",
-    "react-router": "^6.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "5.2.0",
     "react-scripts": "5.0.1",
-    "react-table": "^7.8.0",
-    "typescript": "^4.6.4",
+    "typescript": "4.0.3",
     "web-vitals": "^2.1.4"
+  },
+  "engines": {
+    "node": "16.10.0",
+    "yarn": "1.22.5"
+  },
+  "resolutions": {
+    "mobx": "^6.0.0",
+    "mobx-react": "6.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
