--- src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-   return (
-      <div className="App">
-         Hello World !
-      </div>
-   );
-}
-export default App;
-
