+++ src/pages/app/App.css
@@ -12,7 +12,3 @@
   font-size: calc(10px + 2vmin);
   color: white;
 }
-
-.App-link {
-  color: #61dafb;
-}
