+++ src/pages/app/App.js
@@ -1,0 +1,19 @@
+import './App.css';
+// https://reactjs.org/
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>
+          SMILE Dashboard Home
+        </h1>
+        <p>
+          Hello world!
+        </p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
